Data Extraction/requirements.txt
Mastodon.py
python-dotenv
hdfs